Moved settings code to settings.go
